Implements modal dialog for creating new forum posts
Adds a complete modal interface that allows users to create new forum posts with title and content fields. The modal opens when clicking the new post button and can be closed via a cancel button.

Removes unused character loading functionality that was not relevant to the forum page.

Moves script tags to the end of the body for better page load performance.